Refactor track routes and enhance API structure
- Introduced a new track route file (`track.ts`) to separate single track operations from the existing tracks route, improving organization and maintainability.
- Updated backend API to handle track-related operations under `/api/track` instead of `/api/tracks`, ensuring a clearer and more consistent endpoint structure.
- Enhanced frontend API calls to align with the new route structure, updating method names and endpoints for fetching, uploading, and managing tracks.
- Improved error handling and response consistency across the API, contributing to a more robust application.
- This refactor streamlines the codebase and enhances the overall user experience when interacting with track-related features.
